+++ gantry/constants.py
@@ -1,17 +1,7 @@
 DEFAULT_IMAGE = "petew/gantry"
-
 ENTRYPOINT = "entrypoint.sh"
-
 GITHUB_TOKEN_SECRET = "GITHUB_TOKEN"
-
-CONDA_ENV_FILE = "environment.yml"
-
-CONDA_ENV_FILE_ALTERNATE = "environment.yaml"
-
-PIP_REQUIREMENTS_FILE = "requirements.txt"
-
 RUNTIME_DIR = "/gantry-runtime"
-
 RESULTS_DIR = "/results"
-
+RUNTIME_DIR = "/gantry-runtime"
 METRICS_FILE = f"{RESULTS_DIR}/metrics.json"